+++ app/Controllers/C_Satuan.php
@@ -81,11 +81,23 @@ class C_Satuan extends BaseController
 
     public function hapus($id)
     {
-        $this->satuan->delete($id);
-        session()->setFlashdata('status', 'Data Satuan berhasil dihapus');
-        return redirect()
-            ->to(base_url('tampil-satuan'))
-            ->with('status_icon', 'success')
-            ->with('status_text', 'Data Berhasil dihapus');
+        $where = [
+            'id_satuan' => $id
+        ];
+        $cek_satuan = count((array) $this->satuan->cek_data_direlasi('material', 'kabel', $where));
+        if ($cek_satuan > 0) {
+            session()->setFlashdata('status', 'Cek data Pada tabel material dan kabel pastikan satuan yang terkait sudah tidak ada');
+            return redirect()
+                ->to(base_url('tampil-satuan'))
+                ->with('status_icon', 'warning')
+                ->with('status_text', 'Data Gagal Dihapus');
+        } else {
+            $this->satuan->delete($id);
+            session()->setFlashdata('status', 'Data Satuan berhasil dihapus');
+            return redirect()
+                ->to(base_url('tampil-satuan'))
+                ->with('status_icon', 'success')
+                ->with('status_text', 'Data Berhasil dihapus');
+        }
     }
 }
